Refactor events
* Add string event type class attribute to all event classes. This will
  be used to identify the event type during handler registration and
  notification.
* All common event types should be created as constants in const.py.
* Event classes should be created in the module where they need to be
  used for notification.
* Update tests.
* Fix default config.
