+++ camacq/const.py
@@ -23,3 +23,19 @@ LOG_LEVEL = 'log_level'
 WELL = 'well'
 CONF_PLUGINS = 'plugins'
 PACKAGE = 'camacq'
+
+# events
+BASE_EVENT = 'base_event'
+CAMACQ_START_EVENT = 'camacq_start_event'
+CAMACQ_STOP_EVENT = 'camacq_stop_event'
+CHANNEL_EVENT = 'channel_event'
+FIELD_EVENT = 'field_event'
+IMAGE_REMOVED_EVENT = 'image_removed_event'
+PLATE_EVENT = 'plate_event'
+SAMPLE_EVENT = 'sample_event'
+SAMPLE_IMAGE_EVENT = 'sample_image_event'
+WELL_EVENT = 'well_event'
+COMMAND_EVENT = 'command_event'
+IMAGE_EVENT = 'image_event'
+START_COMMAND_EVENT = 'start_command_event'
+STOP_COMMAND_EVENT = 'stop_command_event'
