+++ tests/test_event.py
@@ -5,7 +5,7 @@ from camacq import event as event_mod
 def test_event_bus(center):
     """Test register handler, fire event and remove handler."""
     event = event_mod.Event({'test': 2})
-    bus = event_mod.EventBus(center)
+    bus = center.bus
 
     def handler(center, event):
         """Handle event."""
@@ -13,11 +13,11 @@ def test_event_bus(center):
             center.data['test'] = 0
         center.data['test'] += event.data['test']
 
-    assert not bus.event_types
+    assert event_mod.BASE_EVENT not in bus.event_types
 
-    remove = bus.register(event_mod.Event, handler)
+    remove = bus.register(event_mod.BASE_EVENT, handler)
 
-    assert bus.event_types == [event_mod.Event]
+    assert event_mod.BASE_EVENT in bus.event_types
     assert not center.data
 
     bus.notify(event)
